refactor: move ogImage jsx inside generateOgImage
Move ogImage component into generateOgImage.tsx file. Delete unnecessary og assets under src
directory.
